perf: add aria-label attributes to improve accessibility & optimize project structure
