+++ src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import Header from './components/Header/Header';
-import PRList from './components/PRList/PRList';
+import PRList from './containers/PRList/PRList';
 import Footer from './components/Footer/Footer';
 
 function App() {
